Move GIO dependencies to a separate file
