+++ photometric_viewer/utils/gio.py
@@ -1,0 +1,8 @@
+import io
+
+from gi.repository import Gio
+
+
+def gio_file_stream(file: Gio.File):
+    _, contents, _ = file.load_contents()
+    return io.TextIOWrapper(io.BytesIO(contents), encoding="utf-8")
